+++ src/frontend/src/components/NotificationManager/components/CreateNSTreeFailureNotif/index.tsx
@@ -1,5 +1,5 @@
 import { B } from "src/common/styles/basicStyle";
-import { ContactUsLink } from "../../../../common/components/library/data_subview/components/ContactUsLink";
+import { ContactUsLink } from "../ContactUsLink";
 
 export const CreateNSTreeFailureNotif = (): JSX.Element => (
   <>
